tests: fix preseed-core20 test without trusted keys
The preseed test was recently updated to be executed on systems included
in the sru validation. As sru validation is executed with
TRUST_TEST_KEYS=false, the test fails trying to get the account-key
(EAD4DbLxK_kn0gzNCXOs3kd6DeMU3f-L6BEsSEuJGBqCORR0gXkdDxMbOm11mRFu) aftet
the test is configured to use the fake store.