tests: post testing metrics (#15968)
* tests: post testing metrics

This is the first approach used to submit testing metrics for spread
executions.

For this purpose, we measure an empty test to calculate the pure
preparation and restoring time of a project, suite, or task.

The focus is mainly on the repeated parts of the code, where any change
could impact the overall testing time.

* add missing secrets

* Update .github/workflows/ci-spread-times.yaml

Co-authored-by: Katie May <katie.m.may@protonmail.com>

* Update tests/utils/metrics/post-metric

Co-authored-by: Katie May <katie.m.may@protonmail.com>

* Update .github/workflows/ci-spread-times.yaml

Co-authored-by: Katie May <katie.m.may@protonmail.com>

---------

Co-authored-by: Katie May <katie.m.may@protonmail.com>