daemon: reject /v2/logs?n=N where N <= 0
The value is parsed as a signed decimal and passed through several
layers to systemd's journalctl -n argument. There negative values have
no meaning.

Note that journalctl does recognize +N as "the last N" vs plain N as
"the first N lines". We could repurpose negative numbers to have the
alternative meaning if we wanted to.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>