cmd/snap-update-ns: add opt-in logging support (#15976)
* cmd/snap-update-ns: add opt-in logging support

When /run/snapd/ns/snap.$SNAP_NAME.log, for per-snap operations, or
/run/snapd/ns/snap.$SNAP_NAME.user.$UID.log, for per-snap-and-user
operations, file exists and can be written to, then configure it as a
write-only (that is, without truncate) log.

This log captures all the operations performed by snap-confine and
snap-update-ns, in relation to constructing and updating a persisted
mount namespace of a given snap.

Jira: https://warthogs.atlassian.net/browse/SNAPDENG-35498

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* cmd/snap-update-ns: add and adjust debug messages

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* cmd/snap-update-ns: fix typo: command

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* cmd/snap-update-ns: reformat comment

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* cmd/snap-update-ns: explain opt-in logging

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

---------

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>