osutil/vfs/lists: depth-first-search tree iterator
The function returns a seq.Iter-like iterator that allows performing
depth-first search of a tree. Since mounts have complex relationships,
the tree nature one is interested in is a type parameter that provides
the list of children, the child node and the parent pointers.

The function DepthFirstSearch is nearly 1-to-1 translation of the Linux
kernel's next_mnt function, with type safety and iterator interfaces
bolted on. The ability to look at a particular tree aspect of a given
type T is provided by the TreeAspect type which provides functions that
take *T as as argument.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>