osutil/vfs/lists: add HeadlessList
HeadlessList[T] are the classic, doubly-linked, circular lists where
the node is embedded in a larger structure and serves as both the
linkge element and traversal starting point. In contrast to List[T]
which requires a dedicated Node[T] in all list elements but does not
have to itself reside in the same type T, nor does it have to be a
member of the list.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>