many: Snap integrity discovery (#14873)
* multiple: support for detecting and using integrity data for seed snaps

This commit adds support for looking up dm-verity integrity data
for asserted snaps in the seed (ones that have verified assertions
available).

snap-bootstrap can then use these integrity data parameters to call to
the snap integrity API in order to either detect available dm-verity
hash data or generate them. It will then proceed with mounting the
snaps using dm-verity.

* seed/seed20.go: add clarifying comments for lookupIntegrityData errors

* c/snap-bootstrap: add tests for install/recover mode with snaps with integrity data

* c/snap-bootstrap: fix some cross-distro unit tests

hardcoding the data-blocks to 4 seem to fail in open suse as the test
snaps that get generated are bigger. Calculate the data blocks value
using the actual size of a test snap instead.

* cmd/snap-bootstrap: pass verity options to writeSysrootMountUnit

* seed/seed20.go: simplify lookupIntegrityData by returning snapRev from deriveSideInfo

* cmd/snap-bootstrap: redefine mountOptions instead of using the global variable

* seed/seed20.go: minor fixes

* multiple: add NewIntegrityDataParams constructor

* c/snap-bootstrap: stop generating dm-verity data when not found on disk

snap-bootstrap won't generate dm-verity from integrity parameters from
snap-revision assertions, if these are not found anymore.

If dm-verity data are not found for a snap that indicates presence of
integrity data in its snap-revision assertion, snap-bootstrap will throw
an error now rather than continue with dm-verity data generation.

* s/integrity: add doc comment for NewIntegrityDataParams

* seed: change error message to lowercase

* c/snap-bootstrap: refactor fs options type checking

to make it easier to expand with different option types in the future

* c/snap-bootstrap: create helper for getting dm-verity mount options

* seed: removed timing for find-integrity-params

* multiple: replace interface{} with any

* c/snap-bootstrap: apply options for mount regardless of type

* snap/integrity: add test for NewIntegrityDataParams

* multiple: refactor lookup of integrity data

add a new NewIntegrityDataParamsFromRevision which takes a snap
revision as input and will return integrity data if found or an error
in the case where integrity data is not found or is of unsupported type.

The choice to ignore error if verity data is not found for a snap
remains with the caller. Ignoring this error in seed20 means that snaps
that don't have integrity data in their snap-revision assertions will
work.

* snap/integrity: fix/add tests for NewIntegrityDataParamsFromRevision

also remove error for unsupported type of integrity data as it
shouldn't really be possible to end up with a snap-revision assertion
with an unsupported type of integrity data.

* fixup! multiple: refactor lookup of integrity data

* c/snap-bootstrap: add todo to use verity data when run mode support is added

* s/integrity: check error using Equals