snapenv: Include snap application environment state
When running an application of a snap is complicated to get some per-app
state values such as its name, desktop file and common IDs, but these
are actually important to make the application to be able to identify
itself properly in a desktop environment.

As per this, expose the application information as env variables.

Related to: https://bugs.launchpad.net/ubuntu/+source/snapd/+bug/2125222