tests/main/snap-run: use test-snapd-sh-core24 to fully execute cross-distro reexec scenario
The test originally used test-snapd-sh, with its implicit base snap
being 'core'. However, the core snap is being repacked during tests
prepare, to include snapd tooling built on the target host. As a result,
when running the test in a cross-distro with reexec scenario, the
test is not fully exhausting the reexec code paths, as some of the
tooling is in fact native.

Switch to test-snapd-sh-core24, so that the snapd tooling is guaranteed
to come from the snapd snap thus ensuring a better coverage of the
reexec case.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>