tests/main/snap-run: downgrade to test-snapd-sh-core22 to work around sudo setup on Fedora
Use test-snapd-sh-core22 to work around the sudo setup on Fedora which
interacts badly with the way snap run --strace calls strace. Sudo is set
up to register the session with logind, places the child processes in a
new cgroup, thus moving it out of a snap specific cgroup we asked
systemd to allocated. See SNAPDENG-35559 for details.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>