ci: use latest Go toolchain on MacOS
The purpose of the MacOS port of parts of snapd is to support "snap pack" in a
non-native Linux environment. On MacOS, Go is not a part of the system, does
not impact system packages and is usually updated to the latest version by
developers that develop with Go on the platform.

Upstream Go is maintained for the past two releases. For example, at the time
of this writing this includes 1.25 and 1.24.  The moment Go 1.26 will release,
1.24 will cease to be supported. In absence of a third party that provides
additional updates, and there's no such party on MacOS, it's unreasonable to
stay on a fixed old version as that is not representative of real-world usage
on the platform.

As such, drop the version constraint from the GitHub action setting up Go for
MacOS, making it use the latest version.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>