tests/main/snap-run: use test-snapd-sh-core22 to fully execute cross-distro reexec scenario (#16029)
* tests/main/snap-run: use test-snapd-sh-core24 to fully execute cross-distro reexec scenario

The test originally used test-snapd-sh, with its implicit base snap
being 'core'. However, the core snap is being repacked during tests
prepare, to include snapd tooling built on the target host. As a result,
when running the test in a cross-distro with reexec scenario, the
test is not fully exhausting the reexec code paths, as some of the
tooling is in fact native.

Switch to test-snapd-sh-core24, so that the snapd tooling is guaranteed
to come from the snapd snap thus ensuring a better coverage of the
reexec case.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* tests/main/snap-run: downgrade to test-snapd-sh-core22 to work around sudo setup on Fedora

Use test-snapd-sh-core22 to work around the sudo setup on Fedora which
interacts badly with the way snap run --strace calls strace. Sudo is set
up to register the session with logind, places the child processes in a
new cgroup, thus moving it out of a snap specific cgroup we asked
systemd to allocated. See SNAPDENG-35559 for details.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

---------

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>