daemons: Allow user daemons to show its icon in the Dock
Some user daemons, like snapd-desktop-integration, can open a
window in the desktop. This makes a new icon to be shown in the
dock while that window is visible.

Unfortunately, in Wayland it is mandatory for an application to
have a .desktop file to map a binary with the corresponding icon
to be shown in the dock, or the currently running applications
list. Until recently, this worked in snapd-desktop-integration
thanks to the BAMF environment variable automagically set by
snapd when installing .desktop files from a snap. Unfortunately,
after mergin https://github.com/canonical/snapd/pull/15798 ,
this is no longer the case, and the windows created by the
daemon have the "default" icon instead of the specific one
defined in the .desktop file.

The reason for this is that the installed .desktop file has its
Exec entry pointing to /snap/bin/XXXXXX (it is modified by snapd
during installation, so there is nothing we can do neither in
the snapcraft.yaml file, nor in the original .desktop file), but
in the case of daemons, snapd never creates a soft link with the
XXXXX name pointing to /usr/bin/snapd, so the dock detects that
the Exec entry is broken and won't honor the Icon entry.

This patch fixes this by creating a soft link to /usr/bin/false
for any daemon that has a .desktop file whose Exec entry points
to the daemon name, and has the NoDisplay entry set to true.
This ensures that the icons are correctly shown while, at the
same time, avoids showing unneeded entries in the application
launchers, along with preventing the user from manually
launching the daemon by accident.