tests/lib: import plz-run v0.4 (#15988)
* tests/lib: import plz-run v0.3

This bundles the complete repository, without history, as of version v0.3.

We can choose to keep or remove files we don't need. For the purpose of snapd
internals tests we will need the go.{mod,sum} and main.go files only.

Jira: https://warthogs.atlassian.net/browse/SNAPDENG-35502

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* tests/lib: remove redundant plz-run files

Keep the code, go module and the LICENSE.

Snapcraft file is removed and a slightly different version will be
re-introduced later, with test-snapd- prefix.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* tests/lib: update plz-run to version 0.4

This fixes one bug and reduces dependencies to bare minimum.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

---------

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>