tests: define an ubuntu kernel version to use with spread tests (#16065)
* tests: allow to define an ubuntu kernel version to use

This change allows to specify the ubuntu kernel version to be used for
tests. This is an example to use this change.

Then this will be integrated with a new workflow which will be used to
test snapd using the different supported kernels.

SPREAD_UPDATE_UBUNTU_^CRNEL_VERSION=6.2.0-32-generic spread -shell
openstack-ps7:ubuntu-22.04-64:tests/main/abort

* Use = instead of == checking for SPREAD_REBOOT