apparmor: adjust interfaces/profiles to cope with coreutils paths
To support use of both GNU and Rust coreutils paths, replace instances
of hardcoded uses of /{usr/,}bin/ and /{usr/,}sbin/ by the
@{SNAP_COREUTIL_DIRS} variable in interfaces and in the policy
template.

Note that apparmor ships a similar variable called @{coreutil_dirs},
but since this variable might not be available cross-distro or when
snap uses the apparmor deb, we are copying it here as well, with the
SNAP prefix.

There's one special case for chroot:
While in core26 chroot is part of /usr/lib/cargo/bin/coreutils/, in
certain Ubuntu releases (or possibly other distros), the binary could
be part of /{,usr/}sbin/, so this patch keeps the original rule.

Fixes: https://bugs.launchpad.net/bugs/2123870
Signed-off-by: Georgia Garcia <georgia.garcia@canonical.com>