Add accessibility interface
A new interface is required to be able to snap the Orca screen
reader. This interface has to give full access to the accessible
bus managed by the AT-SPI daemon, and also to the sockets inside
the at-spi, at-spi2-XXXXX and speech-dispatcher folders located
at /run/user/YYYYY. Finally, it has to allow any desktop binary
to send messages to the accessibility bus no matter if the
receiver (Orca reader) is unconfined or snapped.

Also, this interface is made dependent on `desktop-legacy`, which
is the one that includes the default accessibility permissions,
so it neither can nor have to be manually defined when used as a
plug, but it's automagically added and connected when the old
`desktop-legacy` interface is used by a snap. In the future this
can change, to make that dependent on the base Core snap, thus
making it automatic only for Core24 or less.

Since this interface gives full access to all the accessibility
data, which includes the current contents of any window or
terminal, it must be marked as SUPERPRIVILEGED.