Add accessibility interface
A new interface is required to be able to snap the Orca screen
reader. This interface has to give full access to the accessible
bus managed by the AT-SPI daemon, and also to the sockets inside
the at-spi, at-spi2-XXXXX and speech-dispatcher folders located
at /run/user/YYYYY. Finally, it has to allow any desktop binary
to send messages to the accessibility bus no matter if the
receiver (Orca reader) is unconfined or snapped.

To maintain backwards compatibility, a dependencies system has
been also added. This allows to make an interface dependent on
another. In this case, this new accessibility interface is made
dependent on desktop-legacy interface, which is the one that
includes the default accessibility permissions. This makes that
any snap that has a plug for desktop-legacy will automagically
have a plug for accessibility interface too.

All the AT-SPI rules have, thus, been fully moved from the
`desktop-legacy` to the new `accessibility` interface. The
`AppArmorPermanentPlug` method is used to configure the
`unconfined` rules, thus unifying them in a single interface.

Since this interface gives full access to all the accessibility
data, which includes the current contents of any window or
terminal, it must be marked as SUPERPRIVILEGED.