tests: allow runing test using kernel pattern
This change allows to run the tests using the last kernel which matches
with a pattern. Also it fixes the error in the kernels workflow
execution by removing the \n at the end of the list (in case it finishes
with \n)

Now the kernels workflow contains a pattern used to determine the kernel
to be used instead of a fixed version.