spread: add amazon-linux-2 to garden
I had to install Python 3 for everything to work at a basic level. In the
future this will become easier with automatic installation of all the packages
required for tests.

Moreover, Amazon Linux 2 does not have the power-state cloud-init plugin, so we
need to tell the machine to shut down using imperative logic.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>