osutil/flock_test: drop fancy non-blocking lock sequence, do the simplest thing
Stop being very fancy in the unit test for checking a non-blocking
TryLock(). Previous round of tweaks improved how we synchronize with
flock, but it's clear now that it is still possible that our lock
attempt hits a point between the time when flock shell command opens the
file and takes the lock (as one needs to have a fd to take a the file
lock).

Rework the test to be less fancy, but more robust.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>