daemon: Add support for graphical user daemons
Currently, any daemon with a user scope will be added to the
user's default target. This is not always what is needed,
because some daemons require a graphical desktop environment to
run properly, like snapd-desktop-integration.

This patch adds an extra option to the daemon-scope parameter,
`graphical-user`. When this option is used instead of `user`,
the daemon will have the WantedBy parameter in the .service file
pointing to `graphical-session.target` instead of
`default.target`, and also will include the same target in both
the `After` and `BindsTo` parameters, to ensure that it is
launched when the desktop is ready, and that is stopped when the
desktop session ends.

Implements specification SD-230.