cmd/libsnap-confine-private/device-cgroup-support: more strict check for snap cgroup (#16075)
* cmd/libsnap-confine-private/snap: add helper for converting security tags to unit names

Add a helper for converting security tags to expected systemd unit names.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* cmd/libsnap-confine-private/device-cgroup-support: more strict check for snap cgroup

Make the check for the current process being in a snap cgroup more
strict, by asserting our cgroup matches the snap we're trying to run.

This ensures that even in devmode, if a snap is trying to start another
snap, we do not accidentally install a device filter on the original
snap's cgroup.

Related: SNAPDENG-35414

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

---------

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>