daemon: wrap confdb write values under "values" key (#16087)
Put the map of values to be written to confdb nested under "values".
This helps with backwards compatibility if we need to add new fields to
the PUT body. It's also more consistent with other parts of the API,
namely on read and other related endpoints currently in spec stage.

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>