tests/lib: add libaa-query-peek library
This library can be preloaded into a process to observe how user-space performs
profile queries with the help of the /sys/kernel/security/apparmor/.access
file.

At present nothing uses this library. I used it as a development aid. In the
future I think it could be pre-loaded into dbus.service upon request, by
running spread with an environment variable.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>