osutil/vfs/lists: add explicit initialization functions
With the introduction of Node.container field, the container pointer needs to
be carefully set, as we can no longer reconstruct it with the offset within the
parent. As such the lazy initialization that has carried us thus far, for
initializing next/prev pointers, becomes insufficient for constructing
well-behaving node values.

Add a pair of functions to initialize the container pointer using the
NodePointerer interface. This way the caller cannot misuse the code with
explicit pointer assignment to a subtly wrong value.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

osutil/vfs/lists: add a way to initialize lists

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>