clang-format: specify PointerAlignment
Apparently recent versions of clang-format stopped using
DerivePointerAlignment (which auto-guessed how the pointer is aligned)
with Google style and now defaults to an explicit pointer
alignment (PointerAlignment: Left), which is consistent with similar
recent changes in Google's own style guidelines.

However, most our code is formatted with pointer on the right, so let's
keep that.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>