o/hookstate: only confdb hooks wait if snaps are disabled
Blocking all hooks for unlinked snaps prevents the removal of disabled
snaps because the run-hook task for a "remove" hook hangs the entire
change. This change undoes that blocking for non-confdb hooks since
those typically run as part of a change that already runs the hook after
the snap has been linked or unlinked. Note that the behaviour that this
restores is also wrong since the "remove" hook will just silently fail
to run when removing a disabled snap. I've captured that status quo
along with possible approaches to a complete solution in a spread test.

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>