o/snapstate: InstallPathMany skips refresh control
Installing local snaps should bypass refresh control validation checks
by design. While this is currently true for the single snap install it
wasn't for the *PathMany case because it shares lots of its code with
UpdateMany. This change moves the checks into the store-specific
toUpdate but another option to fix this would be set the
IgnoreValidation flag to bypass the check.

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>