many: allow access to nvidia graphics libraries exported by snaps (#15683)
* snap: and ancillary method to expand a slice containing $SNAP* variables

* dirs: add method returning export folder under /var/lib/snapd/

* interfaces/egl-driver-libs: create configuration for snap-confine

Create configuration files to be read by snap-confine when starting snaps
connected to the opengl interface on classic.

* snap/snapenv: remove /var/lib/snapd/void from SNAP_LIBRARY_PATH

The reason to have it does not apply anymore as we already have two paths in
the variable (see
https://github.com/snapcore/snapd/pull/2732#pullrequestreview-18827193).

* snap/snapenv: set SNAP_LIBRARY_PATH from source files

If there are *.source files in /var/lib/snapd/export/, put those paths in the
SNAP_LIBRARY_PATH variable instead of the using the gl/gl32 paths.

* cmd/snap-confine: mount exported sources

Mount paths exported by nvidia userspace interfaces to the snap mountspace on
classic. This is done to keep compatibility with snaps using the opengl
interface, where we used to mount system libraries in the snap mountspace. Now,
however, these libraries could come from another snaps, so we need to use those
when found.

* interfaces/opengl: allow accessing libs exported by snaps

* interfaces/*-driver-libs: add files with exported paths

Create files that include paths in "source" fields of *-driver-libs interfaces.
These are used by snap-confine so the content of these paths is accessible by
snaps on classic.

* cmd/snap-confine: introduce variable to ease testing of nvidia bits

* tests/nvidia-userspace-libs: check that snaps in classic can use

graphics libraries exported by other snaps.

* cmd/libsnap-confine-private: introduce sc_str_chomp

sc_str_chomp replaces '\n' with '\0' if present at the end of the string.

* tests/snap-env: /var/lib/snapd/void is not in SNAP_LIBRARY_PATH anymore

* interfaces/builtin: minor code improvements

* cmd/snap-confine: copy EGL and Vulkan configuration files

We need to copy instead of symlinking as the files are shipped in snaps to
which we won't have access.

* tests/nvidia-userspace-libs: adapt to changes

* interfaces/builtin: use more generic dirs in some tests

* many: put exported libs in /var/lib/snapd/lib/system/gpu

* interfaces,snap: fix cross-distro tests

* cmd/snap-confine: add vulkan-driver-libs interface to the filter

* interfaces/builtin: add helper returning library-source paths

* snap/snapenv: filter interfaces when creating SNAP_LIBRARY_PATH

as we do in snap-confine.

* interfaces/builtin: use dirs.SnapMountDir instead of dirs.GlobalRootDir in

*-driver-libs tests

* many: add system_ prefix to files under export/

The system_ prefix will give us more control for filtering when we start to use
the export/ directory to allow sharing libraries between snaps.

* interfaces/builtin: rename some helpers to include system prefix