cmd/snap: new tracing mechanism with independently running strace and with shim synchronization
Implement a new tracing mechanism which aims to run strace independently
of the snap application execution pipeline. This has significant
advantage that the snap process runs as the current user, preserving
their permissions. At the same the tracing pipeline, composed of sudo
and strace, is independent from any issues we observed previously, such
as sudo dropping environment variables or executing in a new logind
session (thus a new cgroup).

The mechanism comes with a bit of additional complexity. There is a new
snap-strace-shim helper, which serves as an intermediate synchronization
point where the `snap run` handler knows when to launch strace and wait
for it to attach.

Related: LP#2127766 SNAPDENG-35773 SNAPDENG-35559

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>