confdb: check filters consistent with path and schema
This adds some consistency checks for field filters. First, the accessor
that comes after a field filter must be a map key and not an array index
(e.g., foo[.bar={bar}][{n}] doesn't make sense). Second, if a path ends
with a field filter, then the schema expects it to lead to a map (if it
constrains the data at all).
This also refactor the view rule creation slightly for simplicity and so
we store the original storage path as it was written since we use it for
error messages and any divergence would confuse users.

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>