o/snapstate: InstallPathMany skips refresh control (#16163)
* o/snapstate: InstallPathMany skips refresh control

Installing local snaps should bypass refresh control validation checks
by design. While this is currently true for the single snap install it
wasn't for the *PathMany case because it shares lots of its code with
UpdateMany. This change moves the checks into the store-specific
method on the goal which is a no-op for path installs.

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>