i/p/requestrules: do not record notices for unchanged rules on startup (#16127)
* i/p/requestrules: do not record notices for unchanged rules on startup

Previously, during snapd startup with prompting enabled, the
requestrules backend would record a notice for every rule after loading
rules from disk, regardless of whether the rule had expired or been
changed in any way. This results in all notices being re-written to
disk for every existing rule, of which there could be (very) many.

This is not necessary. For example, the Security Center currently
re-queries all rules whenever it receives a rule notice. It does not
store state and just update the rule with the ID included in the notice.
Therefore, there is no need for snapd to re-record a notice for each
rule on startup.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* i/prompting: record notice if rule is partially expired on load

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* i/prompting: use permission expiration status enum instead of two bools

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* i/prompting: add doc comment about PermExpirationStatus

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* tests: update test to only expect notices for changed rules on load

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* tests: fix MATCH statements in apparmor-prompting-snapd-startup

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* i/p/requestrules: save rules to disk on load if any rule is partially expired

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* i/p/requestrules: add test of partial rule expiration on load

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

---------

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>