httputil/retry.go: consider timeout as no network
We cannot really guess if there is no internet access
available. Having "network" does not mean that network is connected to
the same network as the store. In the case DNS resolves, but there is
a firewall, or connection is just down.

This is an issue with spread tests, where we reset the proxy
configuration and then restart. Snapd is supposed to re-request
for serial, but does not, because it did not detect missing network.