many: reintroduce fdstore helpers
This commit reintroduces reverted PR #16119

* systemd: add fdstore helpers

The initial use case of using systemd's fdstore
is to keep sensitive data that cannot be persisted
to disk like recovery keys and passphrases while
having them survive snapd restarts to increase the
robustness of FDE operation.

* netutil: only use activation sockets passed from systemd

In upcoming work, more fds will be passed from systemd on
startup, previously the helper from go-system considered
all passed fds as activation fds which will not be true
in the future.

* packaging: remove go-systemd dependency

This was the only usage of go-systemd, so I am removing
its dependency as well.

---------

Signed-off-by: Zeyad Gouda <zeyad.gouda@canonical.com>