core-initrd: sort firmware and modules to go to the end
When analyzing deltas for initrd between version, it seems that the
most common case where the delta was big due to compression of the
cpio stream alone is when modules only were modified.

Since initrd is toward the end of EFI or FIT binary, this means
putting modules at the end will make it go to the end of the file and
not mess with double compression due to squashfs compression.