confdb: remove empty parent when removing last leaf
Previously, when removing the last node in a container (map or list) the
container itself wouldn't be removed and would instead be left empty in
the databag. This was inherited from options because we opted to keep
the behaviour of confdb similar unless there was a reason to differ.
We're now introducing filtering on reads and if we we filter a map/list
to the point where none of the data matches the constraints, I think
it's fair to expect a "no data" error instead of an empty container, so
we should make the non-filtered behaviour consistent with that. We've
also introduced --default which relies on this error so an empty
container is unhelpful there.

Besides changing the unset logic to not return/leave empty containers,
this also changes the get logic because it's possible for a read that
goes into a container not to return any data even though the container
isn't empty (see TestGetSeveralUnmatchedPlaceholders for an example).
The only slight oddness here is that we don't prevent users from writing
empty containers so it's possibly to set something to "{}" (for example)
and then read back a NoData error but I think that's more consistent
than the current status quo.

This also improves test coverage for some other corner cases.

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>