c/assemblestate: explicitly close idle connections from http clients (#16206)
* httputil: allow calling CloseIdleConnections on the client returned by NewHTTPClient

* c/assemblestate: explicitly close idle connections from http clients

When running assemblestate_test.TestRun with go1.18, it would sometimes
be very slow due to the process exhausting all available file
descriptors. This does not happen on later go versions because the go
runtime sets RLIMIT_NOFILE to the max allowed limit since go1.19
(https://go.dev/doc/go1.19#runtime).

This is a temp fix, and in reality we should change how we rate limit
outgoing messages. Currently, we rate limit the stream of bytes that we
output, but we don't consider how many concurrent clients there are. So,
a bunch of really small messages can still be sent concurrently.

This problem is also exacerbated by the fact that this test creates many
nodes in the same process, greatly increasing the number of open
connections from the OS's perspective.