cmd/libsnap-confine-private/utils: helper for checking whether a path is 'canonicalized'
Add a helper for checking whether a path is canonicalized, in the sense that it
starts with /, does not end with a trailing / and contains no relative
path elements, such as '.' or '..'. Note that in contrast to realpath(),
the check is only lexical, no symbolic are resolved.

Related: SNAPDENG-35896 sbc#1248682

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>