interfaces/builtin/fwupd: handle case variations in EFI/BOOT/BOOT*.EFI pattern
The UEFI spec outlines a general convention for the name and uses upper
case examples, but also mentions that the storage is on VFAT which is
case insensitive. Apparently there are various variations of the letter
casing in the path out in the wild. We should be able to handle all of
them.

Fixes: https://github.com/fwupd/fwupd/issues/9240

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>