cmd/snap, osutil/strace: new tracing mechanism with independently running strace and with shim synchronization (#16198)
* osutil/strace: support for running strace as a standalone process

Support for running strace as a standalone process which attaches to a
selected tracee.

Related: LP#2127766 SNAPDENG-35773

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* cmd/snap: do not pass -E to sudo when invoking gdbserver

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* tests/main/snap-run: enable on Ubuntu 25.10

Enable the test on Ubuntu 25.10.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* tests/main/snap-run: use core24 based task so that device cgroup is mandatory

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* cmd/snap-strace-shim: tiny helper for synchronization with a tracer program

Add a tiny helper, with the purpuse of providing a synchronization point
for attaching a tracer, right before we exec into the snap-confine ->
snap-exec -> snap app chain.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* packaging: account for snap-strace-shim

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* cmd/snap: new tracing mechanism with independently running strace and with shim synchronization

Implement a new tracing mechanism which aims to run strace independently
of the snap application execution pipeline. This has significant
advantage that the snap process runs as the current user, preserving
their permissions. At the same the tracing pipeline, composed of sudo
and strace, is independent from any issues we observed previously, such
as sudo dropping environment variables or executing in a new logind
session (thus a new cgroup).

The mechanism comes with a bit of additional complexity. There is a new
snap-strace-shim helper, which serves as an intermediate synchronization
point where the `snap run` handler knows when to launch strace and wait
for it to attach.

Related: LP#2127766 SNAPDENG-35773 SNAPDENG-35559

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* tests/main/snap-run: update the test

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

---------

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>