interfaces/builtin/fwupd: wider access to /boot/efi/EFI (#16244)
The UEFI spec outlines a general convention for the name and uses upper
case examples, but also mentions that the storage is on VFAT which is
case insensitive. Apparently there are various variations of the letter
casing in the path out in the wild, making both
/boot/efi/EFI/Boot/Bootx64.efi as well as /boot/efi/EFI/BOOT/BOOTX64.efi
as well as other 'cased' variants correct.

Since the interface is already privileged and needs vetting, and fwupd
essentially needs to be able to read files under /boot/efi/EFI, it makes
sense to give a wider read access.

Fixes: https://github.com/fwupd/fwupd/issues/9240

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>