confdb: support parsing field filters (#16190)
* confdb: refactor databag error usage

The PathError no longer makes sense since the only time we handle it
explicitly in confdb.go is when there's no data to read and there's
already a more specific error for that. The other cases are just generic
errors during traversal which we don't handle and don't need a type.

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>

* confdb: support field filters in paths

Support parsing field filters in view paths, such as
"foo[.field={filter}].bar".

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>

* confdb: check filters consistent with path and schema

This adds some consistency checks for field filters. First, the accessor
that comes after a field filter must be a map key and not an array index
(e.g., foo[.bar={bar}][{n}] doesn't make sense). Second, if a path ends
with a field filter, then the schema expects it to lead to a map (if it
constrains the data at all).
This also refactor the view rule creation slightly for simplicity and so
we store the original storage path as it was written since we use it for
error messages and any divergence would confuse users.

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>

* confdb: cover unexpected termination case

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>

* confdb: rename struct; clarify comments

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>

* confdb: fix comment

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>

* confdb: validate field/filter subkey regex

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>

---------

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>