o/snapstate: avoid potential deadlock when change is aborted (#16235)
* o/snapstate: avoid potential deadlock when change is aborted

If tasks were injected after the change was aborted, it results in a
deadlock between the tasks that would wait for the parent to finish and
the parent that would wait for the children to undo.

Co-authored-by: Dmitry Lyfar <dmitry.lyfar@canonical.com>
Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* o/snapstate: fix InjectTasks unit test for aborted change

chg.Abort() aborts not only its tasks but also
tasks that do not belong to the change but wait
for those that do.

Co-authored-by: Dmitry Lyfar <dmitry.lyfar@canonical.com>
Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

---------

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>
Co-authored-by: Dmitry Lyfar <dmitry.lyfar@canonical.com>