o/i/apparmorprompting: add multiple notices to prompting backend
Add a new method to the prompting notice backend which takes a slice of
notice information and then records notices in a batch, only releasing
the notice backend lock and saving the state to disk once after all
notices have been recorded.

If there is an error saving the state to disk, then all changes are
rolled back. If there is an error when recording one or more notices in
a batch, then any other notices in the batch continue to be recorded.
Such an error should only occur if there is an existing notice with the
same ID but a different user, which should not happen in the real world.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>