Added comment about positional parameters
When a snap is run, the FullName positional is filled only when
using `run snap XXXXX`, but not when using the soft links at
`/snap/bin`. This is the reason why the Execute method in cmd_run
has to check if the FullName element is filled and, if not, use
the normal arguments to know which application to launch.

This patch adds a comment explaining this.