confdb: support field filtering on reads
This includes the core logic for filtering read data based on
constraints applied to field filters. This doesn't include filtering of
unmatched sub-key placeholders or any of the logic required to retain
the constraints during the change. It also only support constraining
string field. While it can be expanded to cover other scalars, we
haven't discussed how to handle typed constraints passed to the CLI
(assume string unless user passes -t as well?) so I opted not to make
assumptions. I'll revisit the issue once we've landed more of this and
test it easily.

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>