gsettings: Do not provide direct access to the dconf database
The gsettings interface is misleading: it promises access to the user
settings via GSettings (that is a GLib API), but it rather provides
access to the dconf API and also its database.

Now, while this should likely be renamed "dconf", since gsettings per se
can work with multiple settings databases, I guess we're too late for
that.

At the same time, given that this API meants to provide *gsettings*
access to the user settings, it should *not* provide access to the
database itself.

As per this, block the direct access to the user settings database
files, while still keep providing access to the dbus API.

That implies that if a snap is connected to this interface:
 - It will still be able to use `gsettings` tool or the `g_settings_*`
   glib API
 - It will be not allowed to directly use the `dconf` tool or its C-API
   counterparts.

I think this is fine since the interface never promised low-level access
but it rather allowed to potentially use the settings through a DBus API
(gsettings).

This is a first step to improve the gsettings situation, as it would
potentially allow us to:
 - Drop the gsettings auto-connection (and make apps to rely on local
   snap databases instead)
 - Add check to the `dconf-daemon` to only allow applications to access
   to the gsettings namespace they claimed to have access to