many: do not allow timeouts < 1us in systemd units
Do not allow timeouts smaller than 1 microsecond in systemd unit options as
those are not allowed for timeouts. Replace instead with a 1us value and log a
warning. We do not return an error to keep compatibility with the old behavior.

Additionally, print units so more granularity than seconds is actually allowed.
Golang units are a subset of those understood by systemd so this is ok.