Reapply "many: re-enable prompting notice backend (#16194)" (revert #16287)
This reverts commit d870ff4f721ed498282d35306633b9acdc084fac.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

o/i/apparmorprompting: handle edge case when creating notice backend

Handle the edge case in which there is a notice which was previously
recorded in snapd state re-recorded in the new notice backend while the
rule or prompt DB is being initialized, prior to the notice backend
being registered as a notice backend. In this case, we want to preserve
the "fresher" data, that is the data of the notice created in the
prompting notice backend, rather than overwrite it with the notice data
drained from the state.

This is truly rare, and can only occur during the the already rare edge
case that there are prompts or rules with active notices at the time
when snapd first restarts after refreshing for the first time to a
version which supports the prompting notice backends. Given that this is
the case, this particular edge case only occurs if a rule expired before
the rules backend was loaded during startup, thus causing a notice that
the rule expired. Now we'll preserve that `{"removed": "expired"}` data.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>