overlord/snapstate: fix inhibition hint being left behind on failed unlink-current-snap
If unlink-current-snap fails after taking the inhibition lock, the hint
file is not cleaned up. This leads to a situation in which `snap run`
refuses to run the snap claiming it is inhibited for refresh.

Add a cleanup, where the inhibition is unlocked on error. Extend the
unit test to make it more realistic and check the scenario.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>