confdb: support placeholder sub-key filtering
Add support for filtering reads based on constraints applied to
unmatched placeholders in the path. The approach taken is similar to the
one used for field filters. It doesn't "fill" the placeholder on
matching like providing a request path does because matched prefix
affects the end result. It also doesn't replace placeholder accessors in
the storage path by concrete key accessors because it creates a
disconnect between the request/storage paths that make the merging of
results from multiple matched rules difficult.

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>