interfaces/builtin: add devlxd interface implementation
This implements an interface for the socker LXD provides to all
instances to allow communication with the hypervisor for additional
configuration (e.g. user supplied metadata). See
https://documentation.ubuntu.com/lxd/latest/dev-lxd/ for details on the
API itself.

The interface slot is hosted on the core snap and auto connection is
denied by default. It allows access to both /dev/lxd/socket and
/var/snap/lxd/common/lxd/devlxd/sock as the first is what we have inside
VM and the second in container instances.