daemon: Add implicit graphical user daemons
This is an implicit implementation of graphical user daemons.
Here, if an user-scoped daemon has the Wayland or the X11 plugs
defined, it is presumed that it requires access to the desktop
graphical environment, so it will depend on graphical-session.target
instead of the default target.

It implements the implicit way of SD-230.

This PR is an alternative to #16091