desktop: Actually use desktop file IDs as desktop IDs (#15811)
* desktop: Actually use desktop file IDs as desktop IDs

The desktop interfaces now exposes the desktop-file-ids, however it does
not takes desktop IDs as parameters.

In fact the specification [1] clearly states that the desktop file ID
must contain the .desktop suffix.

In particular:

  To determine the ID of a desktop file, make its full path relative to
  the $XDG_DATA_DIRS component in which the desktop file is installed,
  remove the "applications/" prefix, and turn '/' into '-'.

  For example /usr/share/applications/foo/bar.desktop has the desktop
  file ID foo-bar.desktop.

As per this, expect the list to contain desktop IDs with .desktop
suffixes by default, and adjust them in case this fails.

[1] https://specifications.freedesktop.org/desktop-entry-spec/latest/file-naming.html#desktop-file-id

* tests: Add test for deprecated fallback desktop id names

Check that desktop IDs without .desktop suffixes are still properly
handled

* fixup! tests: Add test for deprecated fallback desktop id names